requirements.txt
Django==1.11.15
Markdown==2.6.6
amqp==2.3.2
celery==4.1.1
cssmin==0.2.0
django-pipeline==1.6.8
djangorestframework==3.7.7
docutils==0.12
jsmin==2.0.11
pillow==3.4.0
psycopg2==2.7.1
python-memcached==1.57
wsgiref==0.1.2
django-suit==0.2.26
six>=1.10.0
django-axes==2.3.3
requests==2.18.4
coreapi==2.3.3
